fix(distributed): ghost atoms retained across domain boundaries
